+++ src/main/java/wooteco/subway/exception/badrequest/BadRequest.java
@@ -1,0 +1,15 @@
+package wooteco.subway.exception.badrequest;
+
+public class BadRequest extends RuntimeException {
+
+    public BadRequest() {
+    }
+
+    public BadRequest(String message) {
+        super(message);
+    }
+
+    public BadRequest(String message, Throwable cause) {
+        super(message, cause);
+    }
+}
