+++ src/main/java/wooteco/subway/exception/badrequest/StationNotFoundException.java
@@ -1,9 +1,9 @@
-package wooteco.subway.exception;
+package wooteco.subway.exception.badrequest;
 
-public class StationNotFoundException extends SubwayException {
+public class StationNotFoundException extends BadRequest {
     private static final String MESSAGE = "존재하지 않는 지하철 역 입니다.";
-
     public StationNotFoundException() {
+        super(MESSAGE);
     }
 
     public StationNotFoundException(final String message) {
