+++ src/main/java/wooteco/subway/line/infra/line/JdbcLineDao.java
@@ -7,9 +7,9 @@ import org.springframework.jdbc.core.RowMapper;
 import org.springframework.jdbc.support.GeneratedKeyHolder;
 import org.springframework.jdbc.support.KeyHolder;
 import org.springframework.stereotype.Component;
-import wooteco.subway.exception.DuplicatedNameException;
-import wooteco.subway.exception.LineNotFoundException;
-import wooteco.subway.exception.NoRowAffectedException;
+import wooteco.subway.exception.badrequest.DuplicatedNameException;
+import wooteco.subway.exception.badrequest.LineNotFoundException;
+import wooteco.subway.exception.badrequest.NoRowAffectedException;
 import wooteco.subway.line.domain.Line;
 import wooteco.subway.line.domain.section.Distance;
 import wooteco.subway.line.domain.section.Section;
